fix: Fix and remove unused imports
